distance, refactor

+++ src/resources.rs
@@ -2,5 +2,5 @@ use bevy::prelude::Resource;
 
 #[derive(Resource)]
 pub struct WorldRes {
-	pub sun_mass: f32,
+	pub sun_mass: f64,
 }
